// src/types/copy.ts
export interface RsaCopy {
    headlines: string[]
    descriptions: string[]
    paths: string[]
}

export interface PmaxCopy {
    shortHeadlines: string[]
    longHeadlines: string[]
    descriptions: string[]
    businessName: string
}

export interface MetaCopy {
    primaryTexts: string[]
    headlines: string[]
    descriptions: string[]
}